Remove deprecated EventTransformer API (close #68)
