--- src/main/scala/com.snowplowanalytics.snowplow.analytics.scalasdk/json/BooleanFormatException.scala
@@ -1,3 +1,0 @@
-package com.snowplowanalytics.snowplow.analytics.scalasdk.json
-
-class BooleanFormatException(str: String) extends IllegalArgumentException(str)
